% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{algoritmo_probit}
\alias{algoritmo_probit}
\title{Algoritmo de MH para gerar amostras da posteriori do modelo fatorial categ처rico ordenado probit}
\usage{
algoritmo_probit(
  y,
  nit,
  beta,
  f,
  sigma2,
  alpha,
  C0 = 10000,
  a = 0.001,
  b = 0.001,
  sdpropbeta = 0.05,
  sdpropbeta2 = 0.075,
  sdpropf = 0.4,
  sdpropsigma2 = 0.05,
  display_progress = TRUE
)
}
\arguments{
\item{y}{[P x N] matriz de dados completa.}

\item{nit}{[1 x 1] numero de interacoes}

\item{beta}{[P x Q] valor inicial da matriz de cargas fatoriais.}

\item{f}{[Q x N] valor inicial da matriz de fatores.}

\item{sigma2}{[P x 1] valor inicial do vetor de vari창ncias.}

\item{alpha}{[P x K + 1] matriz com os valores fixados dos pontos de corte.}

\item{C0}{[1 x 1] vari창ncia das prioris para beta}

\item{a, b}{[1 x 1] parametros de shape e scale da priori de sigma2}

\item{sdpropbeta, sdpropbeta2, sdpropf, sdpropsigma2}{desvio padrao das propostas}

\item{display_progress}{TRUE para exibir o indicador de progresso.}
}
\description{
Algoritmo de MH para gerar amostras da posteriori do modelo fatorial categ처rico ordenado probit
}
