% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{delta_ijk_probit}
\alias{delta_ijk_probit}
\title{P(Y_ij <= k) - Probabilidade da i-ésima observação da j-ésima variável estar abaixo da categoria k. (PROBIT)}
\usage{
delta_ijk_probit(alpha, beta, f, sigma2)
}
\arguments{
\item{alpha}{[1 x 1] escalar com o ponto de corte associado a k-ésima categoria.}

\item{beta}{[1 x Q] vetor linha com os betas que associam a j-esima variavel e cada um dos Q fatores.}

\item{f}{[Q x 1] vetor coluna com os Q fatores associados a i-ésima observacao.}

\item{sigma2}{[1 x 1] escalar com a variância associada à j-ésima variável.}
}
\value{
[1 x 1] ecalar com a probabilidade de y_ij cair abaixo da categoria k.
}
\description{
P(Y_ij <= k) - Probabilidade da i-ésima observação da j-ésima variável estar abaixo da categoria k. (PROBIT)
}
