% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fatcat-package.R
\docType{package}
\name{fatcat}
\alias{fatcat}
\title{An implementation of the factor model for categorical data}
\description{
The fatcat package provides an implementation of the factor model for categorical data from
[Capdeville, Gonçalves and Pereira (2021)]( https://doi.org/10.1080/02664763.2020.1796935)
}
