% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitfatcat.R
\name{fitfatcat}
\alias{fitfatcat}
\title{Função para gerar amostras da posterior, toda em C++.}
\usage{
fitfatcat(
  y,
  p,
  nit,
  burnin = nit * 0.05,
  lag = 1,
  init_beta = 0.5,
  init_sigma2 = 2,
  sdpropbeta = 0.05,
  sdpropbeta2 = 0.075,
  sdpropf = 0.4,
  sdpropsigma2 = 0.075,
  dist = c("probit", "logit"),
  quiet = F,
  alfa = NULL
)
}
\arguments{
\item{y}{dados. deve ter dimensao jxn}

\item{p}{numero de fatores a serem usados no ajuste}

\item{nit}{numero de interacoes}

\item{burnin}{numero de interacoes iniciais a serem ignoradas}

\item{lag}{pegar uma a cada `lag` interacoes, após burnin.}

\item{init_beta}{valor inicial de beta}

\item{init_sigma2}{valor inicial de sigma2}

\item{sdpropbeta}{desvio padrao da proposta}

\item{sdpropbeta2}{desvio padrao da proposta}

\item{sdpropf}{desvio padrao da proposta}

\item{sdpropsigma2}{desvio padrao da proposta}

\item{dist}{distribuição do erro. atualmente, probit ou logit.}

\item{quiet}{se TRUE, impede o print do indicar de progresso}

\item{alfa}{opcional, fornecer valores verdadeiros de alfa.}
}
\value{
uma lista com as cadeias a posteriori
}
\description{
Função para gerar amostras da posterior, toda em C++.
}
