% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitfatcat.R
\name{fitfatcat_R}
\alias{fitfatcat_R}
\title{Função para gerar amostras da posterior, usando condicionais completas em C++ e
o algoritmo ainda no R. Mantido aqui para criterios de comparaçao de resultado
e performance}
\usage{
fitfatcat_R(
  y,
  p,
  nit,
  init_beta = 0.5,
  init_sigma2 = 2,
  sdpropbeta = 0.05,
  sdpropbeta2 = 0.075,
  sdpropf = 0.4,
  sdpropsigma2 = 0.075
)
}
\arguments{
\item{y}{dados. deve ter dimensao jxn}

\item{p}{numero de fatores a serem usados no ajuste}

\item{nit}{numero de interacoes}

\item{init_beta}{valor inicial de beta}

\item{init_sigma2}{valor inicial de sigma2}

\item{sdpropbeta}{desvio padrao da proposta}

\item{sdpropbeta2}{desvio padrao da proposta}

\item{sdpropf}{desvio padrao da proposta}

\item{sdpropsigma2}{desvio padrao da proposta}
}
\value{
uma lista com as cadeias a posteriori
}
\description{
Função para gerar amostras da posterior, usando condicionais completas em C++ e
o algoritmo ainda no R. Mantido aqui para criterios de comparaçao de resultado
e performance
}
