% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{l_priori_beta}
\alias{l_priori_beta}
\title{(log) Priori completa de beta.}
\usage{
l_priori_beta(beta, C0 = 10000)
}
\arguments{
\item{beta}{[P x Q] matriz beta completa.}

\item{C0}{[1 x 1] variancia da priori de beta.}
}
\value{
[1 x 1] escalar com o valor do log da densidade (nucleo)
}
\description{
(log) Priori completa de beta.
}
