% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{l_priori_beta_j}
\alias{l_priori_beta_j}
\title{(log) Priori para beta_j.}
\usage{
l_priori_beta_j(beta, j, C0 = 10000)
}
\arguments{
\item{beta}{[1 x Q] vetor linha com os betas que associam a j-esima variavel e cada um dos Q fatores.}

\item{j}{[1 x 1] escalar indicando qual é o j associado ao beta_j fornecido.}

\item{C0}{[1 x 1] variancia da priori de beta.}
}
\value{
[1 x 1] escalar com o valor do log da densidade (nucleo)
}
\description{
(log) Priori para beta_j.
}
