% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{l_priori_beta_jl}
\alias{l_priori_beta_jl}
\title{(log) Priori para beta_jl}
\usage{
l_priori_beta_jl(beta, j, l, C0 = 10000)
}
\arguments{
\item{beta}{[1 x 1] valor do beta_jl}

\item{j}{[1 x 1] escalar indicando qual é o j associado a beta_jl}

\item{l}{[1 x 1] escalar indicando qual é o l associado a beta_jl}

\item{C0}{[1 x 1] variancia da priori de beta.}
}
\value{
[1 x 1] escalar com o valor do log da densidade (nucleo) de beta_jl
}
\description{
(log) Priori para beta_jl
}
\details{
As prioris para beta são:
beta_jl ~ N(0, C0), se j < l
beta_jl ~ N(0, C0)I(0, +infty), se j == l
beta_jl = 0 se j > l
}
