% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{l_priori_f}
\alias{l_priori_f}
\title{(log) Priori para f}
\usage{
l_priori_f(f)
}
\arguments{
\item{f}{[Q x N] matriz f completa.}
}
\value{
[1 x 1] escalar com o valor do log da densidade (nucleo)
}
\description{
(log) Priori para f
}
