% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{l_priori_f_i}
\alias{l_priori_f_i}
\title{(log) Priori para f_i}
\usage{
l_priori_f_i(f)
}
\arguments{
\item{f}{[Q x 1] vetor coluna com os Q fatores associados a i-ésima observacao.}
}
\value{
[1 x 1] escalar com o valor do log da densidade (nucleo)
}
\description{
(log) Priori para f_i
}
