% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{l_priori_f_li}
\alias{l_priori_f_li}
\title{(log) Priori para f_li}
\usage{
l_priori_f_li(f)
}
\arguments{
\item{f}{[1 x 1] valor de f_li}
}
\value{
[1 x 1] escalar com o valor do log da densidade (nucleo)
}
\description{
(log) Priori para f_li
}
\details{
A priori para f é
f_li ~ N(0, 1)
}
