% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{l_priori_sigma2}
\alias{l_priori_sigma2}
\title{(log) Priori para sigma2}
\usage{
l_priori_sigma2(sigma2, a = 0.001, b = 0.001)
}
\arguments{
\item{sigma2}{[P x 1] vetor com as variancias das p variáveis.}

\item{a}{[1 x 1] parametro de shape da IG.}

\item{b}{[1 x 1] parametro de scale da IG.}
}
\value{
[1 x 1] escalar com o valor do log da densidade (nucleo)
}
\description{
(log) Priori para sigma2
}
