% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{l_priori_sigma2_j}
\alias{l_priori_sigma2_j}
\title{(log) Priori para sigma2_j}
\usage{
l_priori_sigma2_j(sigma2, a = 0.001, b = 0.001)
}
\arguments{
\item{sigma2}{[1 x 1] valor da variancia da j-ésima variavel}

\item{a}{[1 x 1] parametro de shape da IG.}

\item{b}{[1 x 1] parametro de scale da IG.}
}
\value{
[1 x 1] escalar com o valor do log da densidade (nucleo)
}
\description{
(log) Priori para sigma2_j
}
\details{
A priori para sigma2_j é
sigma2_j ~ IG(a, b)
f(sigma2_j) propto sigma2_j^(- a - 1) exp (-b/sigma2_j)
}
