% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{l_vero_i_probit}
\alias{l_vero_i_probit}
\title{P(Y_i = y_i) - (log) Verossimilhanca para a i-esima observacao.}
\usage{
l_vero_i_probit(y, alpha, beta, f, sigma2)
}
\arguments{
\item{y}{[P x 1] vetor com as p variaveis para a i-esima observacao.}

\item{alpha}{[P x (K+1)]  Matriz alpha com os todos os K + 1 pontos de corte. Pontos de
corte podem diferir para diferentes variaveis, embora nao seja usual.}

\item{beta}{[P x Q] matriz beta completa.}

\item{f}{[Q x 1] vetor coluna com os Q fatores associados a i-ésima observacao.}

\item{sigma2}{[P x 1] vetor sigma2 completo.}
}
\value{
[1 x 1]  ecalar com a log verossimilhanca para todas as variaveis da observacao i.
}
\description{
P(Y_i = y_i) - (log) Verossimilhanca para a i-esima observacao.
}
