% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{l_vero_j_probit}
\alias{l_vero_j_probit}
\title{P(Y_j = y_j) - (log) Verossimilhanca para a j ésima variavel.}
\usage{
l_vero_j_probit(y, alpha, beta, f, sigma2)
}
\arguments{
\item{y}{[1 x N] vetor com as n obersavoes para a j-esima variavel.}

\item{alpha}{[1 x (K+1)] vetor com os todos os K + 1 pontos de corte.}

\item{beta}{[1 x Q] vetor linha com os betas que associam a j-esima variavel e cada um dos Q fatores.}

\item{f}{[Q x N] matriz f completa.}

\item{sigma2}{[1 x 1] escalar com a variância associada à j-ésima variável.}
}
\value{
[1 x 1]  ecalar com a log verossimilhanca para todas as observacoes da variavel j.
}
\description{
P(Y_j = y_j) - (log) Verossimilhanca para a j ésima variavel.
}
