% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{l_vero_probit}
\alias{l_vero_probit}
\title{P(Y = y) - (log) Verossimilhanca (completa).}
\usage{
l_vero_probit(y, alpha, beta, f, sigma2)
}
\arguments{
\item{y}{[P x N] matriz de dados completa.}

\item{alpha}{[P x (K+1)] Matriz alpha com os todos os K + 1 pontos de corte. Pontos de
corte podem diferir para diferentes variaveis, embora nao seja usual.}

\item{beta}{[P x Q] matriz beta completa.}

\item{f}{[Q x N] matriz f completa.}

\item{sigma2}{[P x 1] vetor sigma2 completo.}
}
\value{
[1 x 1] ecalar com a log verossimilhanca completa.
}
\description{
P(Y = y) - (log) Verossimilhanca (completa).
}
