% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitfatcat.R
\name{lastInd}
\alias{lastInd}
\title{https://stackoverflow.com/questions/44099454/in-r-how-to-index-only-the-last-dimension-of-an-array-when-you-dont-know-how-m}
\usage{
lastInd(x, n)
}
\description{
https://stackoverflow.com/questions/44099454/in-r-how-to-index-only-the-last-dimension-of-an-array-when-you-dont-know-how-m
}
