% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordered_data.R
\docType{data}
\name{ordered_data}
\alias{ordered_data}
\title{Simulated data from the factor model for categorical data.}
\format{
## `ordered_data`
A list with 4 elements.
\describe{
  \item{y}{A matrix with dimension 5 x 300, with the actual simulated data.}
  \item{true_beta}{A matrix with dimension 5 x 2, with the values of beta used to generate the data.}
  \item{true_sigma}{A vector with 5 elements, with the values of sigma used to generate the data.}
  \item{true_alpha}{A matrix with dimension 5 x 5, with the values of alpha used to generate the data.}
}
}
\usage{
ordered_data
}
\description{
Simulated data from the factor model for categorical data.
}
\keyword{datasets}
