% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{p_ijk_probit}
\alias{p_ijk_probit}
\title{P(Y_ij = k) - Probabilidade da i-ésima observação da j-ésima variável estar abaixo da categoria k. (PROBIT)}
\usage{
p_ijk_probit(alpha, beta, f, sigma2)
}
\arguments{
\item{alpha}{[1 x 2] vetor com os pontos de corte associados a (k-1)-ésima e (k)-ésima categorias.}

\item{beta}{[1 x Q] vetor linha com os betas que associam a j-esima variavel e cada um dos Q fatores.}

\item{f}{[Q x 1] vetor coluna com os Q fatores associados a i-ésima observacao.}

\item{sigma2}{[1 x 1] escalar com a variância associada à j-ésima variável.}
}
\value{
[1 x 1] ecalar com a probabilidade de y_ij ser exatamente igual acategoria k.
}
\description{
P(Y_ij = k) - Probabilidade da i-ésima observação da j-ésima variável estar abaixo da categoria k. (PROBIT)
}
