% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotfatcat}
\alias{plotfatcat}
\title{Função para plotar os resultados do ajuste}
\usage{
plotfatcat(res, true_values = NULL, name = c("beta", "f", "sigma2"))
}
\arguments{
\item{res}{lista contendo os elemntos beta, f e sigma2, assim como exportado pela
função `fitfatcat`}

\item{true_values}{argumento opcional com os valores verdadeiros dos parametros,
para o caso de estudos de simulação.}

\item{name}{qual parametro deve ser plotado, beta, f ou sigma2?}
}
\value{
nada
}
\description{
Função para plotar os resultados do ajuste
}
