% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rj_fitfatcat.R
\name{q_max}
\alias{q_max}
\title{Funcao auxiliar para identificar o maior valor possivel do numero de fatores
para um dado numero de variáveis.}
\usage{
q_max(m)
}
\arguments{
\item{m}{numero de variaveis, nrow(y)}
}
\value{
limete superior para o numero de fatores
}
\description{
Funcao auxiliar para identificar o maior valor possivel do numero de fatores
para um dado numero de variáveis.
}
