% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rj_fitfatcat.R
\name{rjmcmc_fitfatcat}
\alias{rjmcmc_fitfatcat}
\title{Algoritmo de RJMCMC para o modelo fatorial categórico para dados ordenados}
\usage{
rjmcmc_fitfatcat(
  y,
  alpha,
  chains,
  nit_rjmcmc = 500,
  init_q = 2,
  C0 = 10000,
  a = 0.001,
  b = 0.001,
  a_rj = 2,
  b_rj = 1,
  sdpropbeta = 0.05,
  sdpropbeta2 = 0.075,
  sdpropf = 0.4,
  sdpropsigma2 = 0.075,
  ...
)
}
\arguments{
\item{y}{dados. deve ter dimensao jxn}

\item{alpha}{pontos de corte. usualmente fixados no valor verdadeiro
ou na estimativa usando a correlacao policorica.}

\item{chains}{cadeias com amostras da posteriori dos parametros dado p,
para todos os p's possiveis.}

\item{nit_rjmcmc}{numero de iteracoes para o rjmcmc}

\item{init_q}{numero inicial para o numero de fatores}

\item{C0, a, b}{hiperparametros das prioris do MCMC. Devem coincidir com as cadeias geradas em chains.}

\item{a_rj, b_rj}{hiper parametros para a proposta do RJMCMC}

\item{sdpropbeta, sdpropbeta2, sdpropf, sdpropsigma2}{desvio padrao das propostas.}

\item{...}{argumentos extras.}
}
\description{
Algoritmo de RJMCMC para o modelo fatorial categórico para dados ordenados
}
