% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sim.R
\name{data_sim}
\alias{data_sim}
\title{Função para simular dados do modelo categórico}
\usage{
data_sim(beta, Sigma, alfa, f, link = c("probit", "logit"))
}
\arguments{
\item{beta}{matriz jxp}

\item{Sigma}{matriz de variancias jxj}

\item{alfa}{matriz j x (k+1)}

\item{f}{matriz pxn}

\item{link}{probit ou logit}
}
\value{
matriz jxn
}
\description{
Função para simular dados do modelo categórico
}
\examples{
j <- 5
n <- 300
k <- 4
p.real <- 2
beta.real <- matrix(
  c(
    0.99, 0,
    0, 0.99,
    0.90, 0,
    0, 0.90,
    0.5, 0.5
  ),
  nrow = j, ncol = p.real,
  byrow = TRUE
)
Sigma <- diag(c(0.01, 0.05, 0.10, 0.15, 0.20), nrow = j, ncol = j)
alfa.real <- matrix(
  c(-Inf, qnorm(0.4), qnorm(0.75), qnorm(0.9), Inf),
  nrow = j, ncol = (k + 1),
  byrow = TRUE
)
f.real <- t(MASS::mvrnorm(n, rep(0, p.real), diag(1, p.real, p.real)))

y <- data_sim(beta.real, Sigma, alfa.real, f.real, link = "probit")

}
