% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sim.R
\name{data_sim}
\alias{data_sim}
\title{Funcao para simular dados do modelo categorico}
\usage{
data_sim(
  beta,
  sigma,
  alpha,
  n = NULL,
  f = NULL,
  link = c("probit", "logit"),
  marginal = F
)
}
\arguments{
\item{beta}{matriz jxp}

\item{sigma}{vetor com a diagonal da matriz de variancias, tamanho j}

\item{alpha}{matriz j x (k+1)}

\item{n}{numero de observacoes a serem gerados}

\item{f}{matriz pxn}

\item{link}{probit ou logit}

\item{marginal}{indica se deve usar a distribuicao de y|f ou a distribuicao de y.}
}
\value{
matriz jxn
}
\description{
Funcao para simular dados do modelo categorico
}
\examples{
j <- 5
n <- 300
k <- 4
true_p <- 2

true_beta <- matrix(
  c(
    0.99, 0,
    0, 0.99,
    0.90, 0,
    0, 0.90,
    0.5, 0.5
  ),
  nrow = j, ncol = true_p,
  byrow = TRUE
)

true_sigma <- c(0.01, 0.05, 0.10, 0.15, 0.20)

true_alpha <- matrix(
  c(-Inf, qnorm(0.4), qnorm(0.75), qnorm(0.9), Inf),
  nrow = j, ncol = (k + 1),
  byrow = TRUE
)

true_f <- t(
  MASS::mvrnorm(n, rep(0, true_p), diag(1, true_p, true_p))
)

y <- data_sim(beta = true_beta, sigma = true_sigma, alpha = true_alpha, f = true_f, link = "probit")

}
