% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fatcat-package.R
\docType{package}
\name{fatcat}
\alias{fatcat}
\title{fatcat}
\description{
Ajuste de modelos de analise fatorial para dados categoricos
sob o enfoque bayesiano.
}
\author{
Vitor Capdeville <vgcapdeville@hotmail.com>
}
