% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitfatcat.R
\name{fitfatcat}
\alias{fitfatcat}
\title{Draws samples from the posterior of the factorial model for categorical data.}
\usage{
fitfatcat(
  y,
  q,
  nit,
  burnin = 0,
  lag = 1,
  init_beta = 0.5,
  init_sigma2 = 2,
  C0 = 1e+05,
  a = 0.001,
  b = 0.001,
  sdpropbeta = 0.05,
  sdpropbeta2 = 0.075,
  sdpropf = 0.4,
  sdpropsigma2 = 0.075,
  dist = c("probit"),
  quiet = F,
  alpha = NULL
)
}
\arguments{
\item{y}{A matrix of integers with dimension j x n, where j is the number of variables and n the number of observations.}

\item{q}{An integer with the number of factors.}

\item{nit}{An integer with the number of iterations.}

\item{burnin}{An integer with the number of initial iterations to be ignored.}

\item{lag}{An integer with the lag of the iterations taken. ``lag = 1`` means take every iteration. ``lag = n`` means
take one iteration every n iterations.}

\item{init_beta}{An integer with the initial value of the first column of beta.}

\item{init_sigma2}{An integer with the initial value of sigma2. All elements of the sigma2 vector will be initialized
with the same value.}

\item{C0}{An integer with the variance of the prior for beta.}

\item{a, b}{An integer with the shape and scale parameters of the prior of sigma2.}

\item{sdpropbeta}{An integer with the standard deviation of the proposal of betas that are unrestricted.}

\item{sdpropbeta2}{An integer with the standard deviation of the proposal of betas that are restricted to be greater
than or equal to zero.}

\item{sdpropf}{An integer with the standard deviation of the proposal of factors.}

\item{sdpropsigma2}{An integer with the standard deviation of the proposal of sigma2.}

\item{dist}{An string with the name of the error distribution. Currently, only `probit` is accepted.}

\item{quiet}{If TRUE, prevents the display of progress bar.}

\item{alpha}{An vector with the true value of alpha. Usually, starts with -Inf and ends with +Inf.}
}
\value{
A posterior object with a sample from the posterior distribution.
}
\description{
Draws samples from the posterior of the factorial model for categorical data.
}
