% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitfatcat.R
\name{fitfatcat}
\alias{fitfatcat}
\title{Função para gerar amostras da posterior, toda em C++.}
\usage{
fitfatcat(
  y,
  q,
  nit,
  burnin = 0,
  lag = 1,
  init_beta = 0.5,
  init_sigma2 = 2,
  C0 = 10000,
  a = 0.001,
  b = 0.001,
  sdpropbeta = 0.05,
  sdpropbeta2 = 0.075,
  sdpropf = 0.4,
  sdpropsigma2 = 0.075,
  dist = c("probit", "logit"),
  quiet = F,
  alpha = NULL
)
}
\arguments{
\item{y}{dados. deve ter dimensao jxn}

\item{q}{numero de fatores a serem usados no ajuste}

\item{nit}{numero de interacoes}

\item{burnin}{numero de interacoes iniciais a serem ignoradas}

\item{lag}{pegar uma a cada `lag` interacoes, após burnin.}

\item{init_beta}{valor inicial de beta}

\item{init_sigma2}{valor inicial de sigma2}

\item{C0}{variância das prioris para beta}

\item{a, b}{parametros de shape e scale da priori de sigma2}

\item{sdpropbeta}{desvio padrao da proposta}

\item{sdpropbeta2}{desvio padrao da proposta}

\item{sdpropf}{desvio padrao da proposta}

\item{sdpropsigma2}{desvio padrao da proposta}

\item{dist}{distribuição do erro. atualmente, probit ou logit.}

\item{quiet}{se TRUE, impede o print do indicar de progresso}

\item{alpha}{opcional, fornecer valores verdadeiros de alpha.}
}
\value{
uma lista com as cadeias a posteriori
}
\description{
Função para gerar amostras da posterior, toda em C++.
}
